Add option for saving and restoring simnet initial state

+++ bitcoin/docker/bin/start.sh
@@ -73,24 +73,42 @@ sshpass -p toor ssh -nNT -o StrictHostKeyChecking=no -R ${RPCPORT1EXT}:localhost
 sshpass -p toor ssh -nNT -o StrictHostKeyChecking=no -R ${RPCPORT2EXT}:localhost:${RPCPORT2} root@localhost &
 sshpass -p toor ssh -nNT -o StrictHostKeyChecking=no -R ${RPCPORT3EXT}:localhost:${RPCPORT3} root@localhost &
 
-start_btcd &
-start_lnd /simnet/lnd0 $PEERPORT0 $RPCPORT0 &
-sleep 10
-MINING_ADDRESS=$(lncli --rpcserver localhost:$RPCPORT0 newaddress p2wkh | jq  -r ".address")
-echo "MINING_ADDRESS=" $MINING_ADDRESS
-kill $(pgrep btcd)
-sleep 3
-# Now start btcd again. MINING_ADDRESS is set. So it should use it as a mining address.
-start_btcd &
-sleep 3
-start_btcctl generate 1025
-sleep 10
+if [ "$1" == "--restore-init-state" ]; then
+    cp -a /save_simnet/* /simnet
+    MINING_ADDRESS=$(cat /simnet/MINING_ADDRESS)
+    start_btcd &
+    start_lnd /simnet/lnd0 $PEERPORT0 $RPCPORT0 &
+else
+    start_btcd &
+    start_lnd /simnet/lnd0 $PEERPORT0 $RPCPORT0 &
+    sleep 10
+    MINING_ADDRESS=$(lncli --rpcserver localhost:$RPCPORT0 newaddress p2wkh | jq  -r ".address")
+    echo "MINING_ADDRESS=" $MINING_ADDRESS
+    echo $MINING_ADDRESS>/simnet/MINING_ADDRESS
+    kill $(pgrep btcd)
+    sleep 3
+    # Now start btcd again. MINING_ADDRESS is set. So it should use it as a mining address.
+    start_btcd &
+    sleep 3
+    start_btcctl generate 1025
+    sleep 10
+fi
 
 start_lnd /simnet/lnd1 $PEERPORT1 $RPCPORT1 &
 start_lnd /simnet/lnd2 $PEERPORT2 $RPCPORT2 &
 start_lnd /simnet/lnd3 $PEERPORT3 $RPCPORT3 &
 
 sleep 15
+
+if [ "$1" == "--save-init-state" ]; then
+    echo "Saving simnet state"
+    pkill -SIGINT lnd
+    sleep 5
+    pkill -SIGINT btcd
+    sleep 1
+    cp -a /simnet/btcd /simnet/lnd{0,1,2,3} /simnet/MINING_ADDRESS /save_simnet
+    exit 0
+fi
 
 ADDRNODE1=$(lncli --rpcserver localhost:$RPCPORT1 newaddress p2wkh | jq  -r ".address")
 ADDRNODE2=$(lncli --rpcserver localhost:$RPCPORT2 newaddress p2wkh | jq  -r ".address")
@@ -121,7 +139,7 @@ lncli --rpcserver localhost:$RPCPORT3 walletbalance
 #Connect first node to second and create a channel
 lncli --rpcserver localhost:${RPCPORT1} connect ${IDENTITYKEY2}@localhost:${PEERPORT2}
 sleep 1
-lncli --rpcserver localhost:${RPCPORT1} openchannel --node_key ${IDENTITYKEY2} --local_amt ${CHANNELSIZE} --push_amt ${PUSHAMOUNT} --num_confs 1
+lncli --rpcserver localhost:${RPCPORT1} openchannel --node_key ${IDENTITYKEY2} --local_amt ${CHANNELSIZE} --push_amt ${PUSHAMOUNT}
 sleep 1
 start_btcctl generate 10
 sleep 1
@@ -129,7 +147,7 @@ sleep 1
 #Connect second node to third and create a channel
 lncli --rpcserver localhost:${RPCPORT2} connect ${IDENTITYKEY3}@localhost:${PEERPORT3}
 sleep 1
-lncli --rpcserver localhost:${RPCPORT2} openchannel --node_key ${IDENTITYKEY3} --local_amt ${CHANNELSIZE} --push_amt ${PUSHAMOUNT} --num_confs 1
+lncli --rpcserver localhost:${RPCPORT2} openchannel --node_key ${IDENTITYKEY3} --local_amt ${CHANNELSIZE} --push_amt ${PUSHAMOUNT}
 sleep 1
 start_btcctl generate 10
